// src/app/interfaces/page.model.ts
  export interface productList {
    sNo?: number;
    ProductName?: string;
    SKU?: string;
    Category?: string;
    isSelected?: boolean;
    Brand?: string;
    price?: string;
    Unit?: string;
    Qty?: string;
    CreatedBy?: string;
    img?: string;
  }
// ...
  export interface apiResultFormat {
    data: [];
    totalData: number;
  }